@@ -33,27 +33,17 @@
   },
   "author": "Abe Hendricks<moemanchu@icloud.com>",
   "devDependencies": {
-    "@types/react": "^17.0.4",
-    "@typescript-eslint/eslint-plugin": "^4.6.1",
-    "@typescript-eslint/parser": "^4.6.1",
-    "eslint": "^7.12.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "@types/react": "^17.0.25",
     "install-peers-cli": "^2.2.0",
-    "microbundle": "^0.13.0",
-    "prettier": "^2.0.4",
-    "typescript": "^4.0.5"
+    "microbundle": "^0.13.3",
+    "typescript": "^4.4.3"
   },
   "peerDependencies": {
-    "@material-ui/core": "^4.11.3",
+    "@mui/material": "^5.0.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   },
   "dependencies": {
-    "@material-ui/icons": "^4.11.2",
-    "clsx": "^1.1.1"
+    "@mui/icons-material": "^5.0.1"
   }
 }
