@@ -1,2 +1,2 @@
-import r,{useState as e,useCallback as o,useMemo as t}from"react";import{withStyles as n,createStyles as a,CircularProgress as i,Fab as l}from"@material-ui/core";import c from"@material-ui/icons/GpsFixed";import s from"clsx";import u from"@material-ui/core/colors/green";function m(){return(m=Object.assign||function(r){for(var e=1;e<arguments.length;e++){var o=arguments[e];for(var t in o)Object.prototype.hasOwnProperty.call(o,t)&&(r[t]=o[t])}return r}).apply(this,arguments)}function f(r,e){if(null==r)return{};var o,t,n={},a=Object.keys(r);for(t=0;t<a.length;t++)e.indexOf(o=a[t])>=0||(n[o]=r[o]);return n}var p=n(a({progress:{color:u[500],position:"absolute",top:-4,left:-4,zIndex:1}}))(function(n){var a=n.classes,u=n.timeout,p=void 0===u?15e3:u,g=n.size,v=void 0===g?"small":g,d=n.onClick,b=n.onGeolocateError,x=n.circularProgressProps,E=f(n,["classes","timeout","size","onClick","onGeolocateError","circularProgressProps"]),P=e(!1),h=P[0],y=P[1],z=o(function(){y(!0),navigator.geolocation.getCurrentPosition(function(r){var e=r.coords;y(!1),d(e)},function(r){y(!1),null==b||b(r)},{enableHighAccuracy:!1,maximumAge:0,timeout:p})},[d,b,p]),O=t(function(){var e=null!=x?x:{},o=e.className,t=f(e,["className"]);return h&&r.createElement(i,m({size:48,className:s([a.progress,o])},t))},[h,x,a]);return r.createElement(r.Fragment,null,r.createElement(l,m({size:v,"aria-label":"geolocate control",onClick:z},E),r.createElement(c,null)),O)});export default p;
+import e,{useState as r,useCallback as o,useMemo as t}from"react";import{CircularProgress as n,Fab as i}from"@mui/material";import a from"@mui/icons-material/GpsFixed";function l(){return(l=Object.assign||function(e){for(var r=1;r<arguments.length;r++){var o=arguments[r];for(var t in o)Object.prototype.hasOwnProperty.call(o,t)&&(e[t]=o[t])}return e}).apply(this,arguments)}function c(e,r){if(null==e)return{};var o,t,n={},i=Object.keys(e);for(t=0;t<i.length;t++)r.indexOf(o=i[t])>=0||(n[o]=e[o]);return n}export default function(s){var u=s.timeout,m=void 0===u?15e3:u,f=s.size,p=void 0===f?"small":f,g=s.onClick,v=s.onGeolocateError,x=s.circularProgressProps,d=c(s,["classes","timeout","size","onClick","onGeolocateError","circularProgressProps"]),b=r(!1),y=b[0],E=b[1],P=o(function(){E(!0),navigator.geolocation.getCurrentPosition(function(e){var r=e.coords;E(!1),g(r)},function(e){E(!1),null==v||v(e)},{enableHighAccuracy:!1,maximumAge:0,timeout:m})},[g,v,m]),h=t(function(){var r=null!=x?x:{},o=r.sx,t=c(r,["className","sx"]);return y&&e.createElement(n,l({size:48,sx:l({color:"secondary.main",position:"absolute",top:"-4px",left:"-4px",zIndex:1},o)},t))},[y,x]);return e.createElement(e.Fragment,null,e.createElement(i,l({size:p,"aria-label":"geolocate control",onClick:P},d),e.createElement(a,null)),h)}
 //# sourceMappingURL=index.esm.js.map
