@@ -1,2 +1,2 @@
-var e=require("react"),r=require("@material-ui/core"),t=require("@material-ui/icons/GpsFixed"),a=require("clsx"),o=require("@material-ui/core/colors/green");function l(e){return e&&"object"==typeof e&&"default"in e?e:{default:e}}var n=l(e),i=l(t),u=l(a);function c(){return(c=Object.assign||function(e){for(var r=1;r<arguments.length;r++){var t=arguments[r];for(var a in t)Object.prototype.hasOwnProperty.call(t,a)&&(e[a]=t[a])}return e}).apply(this,arguments)}function s(e,r){if(null==e)return{};var t,a,o={},l=Object.keys(e);for(a=0;a<l.length;a++)r.indexOf(t=l[a])>=0||(o[t]=e[t]);return o}var f=r.createStyles({progress:{color:l(o).default[500],position:"absolute",top:-4,left:-4,zIndex:1}}),m=r.withStyles(f)(function(t){var a=t.classes,o=t.timeout,l=void 0===o?15e3:o,f=t.size,m=void 0===f?"small":f,d=t.onClick,g=t.onGeolocateError,p=t.circularProgressProps,v=s(t,["classes","timeout","size","onClick","onGeolocateError","circularProgressProps"]),b=e.useState(!1),y=b[0],P=b[1],h=e.useCallback(function(){P(!0),navigator.geolocation.getCurrentPosition(function(e){var r=e.coords;P(!1),d(r)},function(e){P(!1),null==g||g(e)},{enableHighAccuracy:!1,maximumAge:0,timeout:l})},[d,g,l]),x=e.useMemo(function(){var e=null!=p?p:{},t=e.className,o=s(e,["className"]);return y&&n.default.createElement(r.CircularProgress,c({size:48,className:u.default([a.progress,t])},o))},[y,p,a]);return n.default.createElement(n.default.Fragment,null,n.default.createElement(r.Fab,c({size:m,"aria-label":"geolocate control",onClick:h},v),n.default.createElement(i.default,null)),x)});module.exports=m;
+var e=require("react"),r=require("@mui/material"),t=require("@mui/icons-material/GpsFixed");function o(e){return e&&"object"==typeof e&&"default"in e?e:{default:e}}var n=o(e),a=o(t);function l(){return(l=Object.assign||function(e){for(var r=1;r<arguments.length;r++){var t=arguments[r];for(var o in t)Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o])}return e}).apply(this,arguments)}function i(e,r){if(null==e)return{};var t,o,n={},a=Object.keys(e);for(o=0;o<a.length;o++)r.indexOf(t=a[o])>=0||(n[t]=e[t]);return n}module.exports=function(t){var o=t.timeout,u=void 0===o?15e3:o,c=t.size,s=void 0===c?"small":c,f=t.onClick,m=t.onGeolocateError,d=t.circularProgressProps,p=i(t,["classes","timeout","size","onClick","onGeolocateError","circularProgressProps"]),g=e.useState(!1),v=g[0],x=g[1],b=e.useCallback(function(){x(!0),navigator.geolocation.getCurrentPosition(function(e){var r=e.coords;x(!1),f(r)},function(e){x(!1),null==m||m(e)},{enableHighAccuracy:!1,maximumAge:0,timeout:u})},[f,m,u]),y=e.useMemo(function(){var e=null!=d?d:{},t=e.sx,o=i(e,["className","sx"]);return v&&n.default.createElement(r.CircularProgress,l({size:48,sx:l({color:"secondary.main",position:"absolute",top:"-4px",left:"-4px",zIndex:1},t)},o))},[v,d]);return n.default.createElement(n.default.Fragment,null,n.default.createElement(r.Fab,l({size:s,"aria-label":"geolocate control",onClick:b},p),n.default.createElement(a.default,null)),y)};
 //# sourceMappingURL=index.js.map
