@@ -1,2 +1,2 @@
-!function(e,r){"object"==typeof exports&&"undefined"!=typeof module?module.exports=r(require("react"),require("@material-ui/core"),require("@material-ui/icons/GpsFixed"),require("clsx"),require("@material-ui/core/colors/green")):"function"==typeof define&&define.amd?define(["react","@material-ui/core","@material-ui/icons/GpsFixed","clsx","@material-ui/core/colors/green"],r):(e||self).reactMuiGeolocator=r(e.react,e.core,e.GpsFixedIcon,e.clsx,e.green)}(this,function(e,r,t,o,a){function l(e){return e&&"object"==typeof e&&"default"in e?e:{default:e}}var n=l(e),i=l(t),c=l(o);function u(){return(u=Object.assign||function(e){for(var r=1;r<arguments.length;r++){var t=arguments[r];for(var o in t)Object.prototype.hasOwnProperty.call(t,o)&&(e[o]=t[o])}return e}).apply(this,arguments)}function s(e,r){if(null==e)return{};var t,o,a={},l=Object.keys(e);for(o=0;o<l.length;o++)r.indexOf(t=l[o])>=0||(a[t]=e[t]);return a}var f=r.createStyles({progress:{color:l(a).default[500],position:"absolute",top:-4,left:-4,zIndex:1}});return r.withStyles(f)(function(t){var o=t.classes,a=t.timeout,l=void 0===a?15e3:a,f=t.size,d=void 0===f?"small":f,m=t.onClick,p=t.onGeolocateError,g=t.circularProgressProps,b=s(t,["classes","timeout","size","onClick","onGeolocateError","circularProgressProps"]),v=e.useState(!1),y=v[0],x=v[1],h=e.useCallback(function(){x(!0),navigator.geolocation.getCurrentPosition(function(e){var r=e.coords;x(!1),m(r)},function(e){x(!1),null==p||p(e)},{enableHighAccuracy:!1,maximumAge:0,timeout:l})},[m,p,l]),P=e.useMemo(function(){var e=null!=g?g:{},t=e.className,a=s(e,["className"]);return y&&n.default.createElement(r.CircularProgress,u({size:48,className:c.default([o.progress,t])},a))},[y,g,o]);return n.default.createElement(n.default.Fragment,null,n.default.createElement(r.Fab,u({size:d,"aria-label":"geolocate control",onClick:h},b),n.default.createElement(i.default,null)),P)})});
+!function(e,t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("react"),require("@mui/material"),require("@mui/icons-material/GpsFixed")):"function"==typeof define&&define.amd?define(["react","@mui/material","@mui/icons-material/GpsFixed"],t):(e||self).reactMuiGeolocator=t(e.react,e.material,e.GpsFixedIcon)}(this,function(e,t,r){function o(e){return e&&"object"==typeof e&&"default"in e?e:{default:e}}var n=o(e),a=o(r);function i(){return(i=Object.assign||function(e){for(var t=1;t<arguments.length;t++){var r=arguments[t];for(var o in r)Object.prototype.hasOwnProperty.call(r,o)&&(e[o]=r[o])}return e}).apply(this,arguments)}function l(e,t){if(null==e)return{};var r,o,n={},a=Object.keys(e);for(o=0;o<a.length;o++)t.indexOf(r=a[o])>=0||(n[r]=e[r]);return n}return function(r){var o=r.timeout,u=void 0===o?15e3:o,c=r.size,s=void 0===c?"small":c,f=r.onClick,d=r.onGeolocateError,m=r.circularProgressProps,p=l(r,["classes","timeout","size","onClick","onGeolocateError","circularProgressProps"]),g=e.useState(!1),x=g[0],b=g[1],v=e.useCallback(function(){b(!0),navigator.geolocation.getCurrentPosition(function(e){var t=e.coords;b(!1),f(t)},function(e){b(!1),null==d||d(e)},{enableHighAccuracy:!1,maximumAge:0,timeout:u})},[f,d,u]),y=e.useMemo(function(){var e=null!=m?m:{},r=e.sx,o=l(e,["className","sx"]);return x&&n.default.createElement(t.CircularProgress,i({size:48,sx:i({color:"secondary.main",position:"absolute",top:"-4px",left:"-4px",zIndex:1},r)},o))},[x,m]);return n.default.createElement(n.default.Fragment,null,n.default.createElement(t.Fab,i({size:s,"aria-label":"geolocate control",onClick:v},p),n.default.createElement(a.default,null)),y)}});
 //# sourceMappingURL=index.umd.js.map
