@@ -1,22 +1,22 @@
-import {createMuiTheme} from '@material-ui/core/styles'
-import red from '@material-ui/core/colors/red'
+import {createTheme} from '@mui/material';
+import {red} from '@mui/material/colors';
 
 // Create a theme instance.
-const theme = createMuiTheme({
+const theme = createTheme({
   palette: {
     primary: {
-      main: '#556cd6'
+      main: '#556cd6',
     },
     secondary: {
-      main: '#19857b'
+      main: '#19857b',
     },
     error: {
-      main: red.A400
+      main: red.A400,
     },
     background: {
-      default: '#fff'
-    }
-  }
-})
+      default: '#fff',
+    },
+  },
+});
 
-export default theme
+export default theme;
